fix(types): export ConfigStore in /types cli-core export sub-paths
