+++ packages/@sanity/cli-core/src/_exports/config.ts
@@ -2,6 +2,7 @@
 
 export {
   clearCliTokenCache,
+  type ConfigStore,
   getCliToken,
   getCliUserConfig,
   getUserConfig,
