+++ packages/@sanity/cli-core/src/_exports/index.ts
@@ -10,7 +10,6 @@ export {
 } from '../apiClient.js'
 export {
   clearCliTokenCache,
-  type ConfigStore,
   getCliToken,
   getCliUserConfig,
   getUserConfig,
@@ -18,7 +17,7 @@ export {
 } from '../config/cli/cliUserConfig.js'
 export {getCliConfig, getCliConfigUncached} from '../config/cli/getCliConfig.js'
 export {getCliConfigSync} from '../config/cli/getCliConfigSync.js'
-export {type CliConfig} from '../config/cli/types/cliConfig.js'
+export {type CliConfig, type ConfigStore} from '../config/cli/types/cliConfig.js'
 export {type UserViteConfig} from '../config/cli/types/userViteConfig.js'
 export {type ApplicationType, isWorkbenchApp} from '../config/cli/workbenchApp.js'
 export {findProjectRoot} from '../config/findProjectRoot.js'
